moved Workflow description from backends._parsl to backends._base
